Flytter funksjoner til ny fil
